+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### v0.0.42
+
+- Add style textAlignVertical:'top' to TextInput when multiline for same behavior in both platforms
+
+--
 ### v0.0.41
 
 - Add Note list item component